Разнесет вспомогательные функции по разным файлам, исправит соответствующие импорты
