+++ src/presenter/trip-events.js
@@ -4,7 +4,8 @@ import LoadingView from '../view/loading';
 import EventPresenter from './event';
 import EmptyEventsListView from '../view/empty-events-list';
 import SortPresenter from './sort';
-import { sortDateDown, sortTimeUp, sortPriceUp, filter } from '../utils';
+import { sortDateDown, sortTimeUp, sortPriceUp } from '../utils/sort-utils.js';
+import { filter } from '../utils/filter-utils.js';
 import { UserAction, UpdateType, FilterType, SortType } from '../const';
 import NewEventPresenter from './new-event-presenter';
 import UiBlocker from '../framework/ui-blocker/ui-blocker.js';
